Refactor: Consolidate game logic and centralize ability data
This commit refactors the project structure for better organization and easier modification of game stats and effects.

Key changes include:

1.  **Entity Consolidation**:
    *   Merged `Entity.js`, `Unit.js`, `Fortress.js`, and `Projectile.js` into a single `entities.js` file.
    *   `EntityManager.js` was moved to the root `IX/` directory.

2.  **Manager Consolidation**:
    *   Merged `AIController.js`, `GridManager.js`, `Player.js`, `PlayerManager.js`, and `WaveManager.js` into a single `managers.js` file.

3.  **System Consolidation**:
    *   Merged `AbilitySystem.js`, `CombatSystem.js`, `MovementSystem.js`, and `UpgradeSystem.js` into a single `gameSystems.js` file.
    *   `VisualEffectSystem.js` and `SystemManager.js` remain in the `IX/systems/` directory.

4.  **Data-Driven Abilities**:
    *   Moved numerous hardcoded parameters for abilities (e.g., heal amounts, damage bonuses, durations, radii, proc chances) from `AbilitySystem.js` (now part of `gameSystems.js`) into `data/unitTypes.js`.
    *   Updated `AbilitySystem` logic to read these parameters from the unit/ability definitions, making stats and effects more configurable.

5.  **UI Consolidation**:
    *   Merged `InputManager.js` into `UIManager.js`.

These changes aim to reduce the number of files, group related logic, and improve the maintainability and configurability of unit abilities without altering game features.
